dog_breed_identifier_backend/requirements.txt
flask==1.1.2
flask-cors==3.0.9
numpy==1.19.2
redis==3.5.3
pillow==8.0.1